add tests, extract to a separate file
