+++ client/lib/postal-code/README.md
@@ -1,0 +1,4 @@
+Postal Code Utils
+======
+
+This module provides utils for handling postal codes.
